requirements.txt
tk
customtkinter
Pillow
pyinstaller
pygame-ce